requirements.txt
torch>=2.0.0
Pillow>=9.0.0
torchvision>=0.15.0
nltk>=3.8
wandb>=0.15.0
tqdm>=4.64.0
transformers>=4.30.0
safetensors>=0.3.0
flash_attn>=2.0.0
huggingface_hub>=0.16.0
accelerate>=0.20.0